+++ front-end/src/pages/Member/hooks/useSearchId.js
@@ -29,8 +29,8 @@ function useSearchId() {
     else setNameError('');
 
     const emailRegex =
-      /([\w-.]+)@((\[[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\.)|(([\w-]+\.)+))([a-zA-Z]{2,4}|[0-9]{1,3})(\]?)$/;
-    if (!emailRegex.test(email))
+      /^[A-Za-z0-9]([-_.]?[A-Za-z0-9])*@[A-Za-z0-9]([-_.]?[A-Za-z0-9])*\.[A-Za-z]{2,3}$/i;
+    if (!emailRegex.test(inputValue.email) || inputValue.email.length < 1)
       setEmailError('올바른 이메일 형식이 아닙니다.');
     else setEmailError('');
